+++ src/Framework/FieldsAPI/Types.php
@@ -4,12 +4,15 @@ namespace Give\Framework\FieldsAPI;
 
 /**
  * @since 2.12.0
+ * @unreleased add Form and Group
  */
 class Types {
 	const CHECKBOX = Checkbox::TYPE;
 	const DATE     = Date::TYPE;
 	const EMAIL    = Email::TYPE;
 	const FILE     = File::TYPE;
+	const FORM     = Form::TYPE;
+	const GROUP    = Group::TYPE;
 	const HIDDEN   = Hidden::TYPE;
 	const PHONE    = Phone::TYPE;
 	const RADIO    = Radio::TYPE;