+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Add missing `TYPE` to Fields API `Group` node type (#5895)
 
+### Changed
+
+-   Do not extend Fields API `Form` from `Group` (#5889)
+
 ## 2.12.1 - 2021-07-22
 
 ### Fixed