+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Add missing `TYPE` to Fields API `Group` node type (#5895)
+-   Add missing Fields API node types to `Types` (#5891)
 
 ## 2.12.1 - 2021-07-22